+++ requirements.txt
@@ -3,11 +3,7 @@ azure-storage-blob==0.37.0
 azure-storage-queue==0.37.0
 boto3==1.4.4
 connexion==1.1.16
-elasticsearch-dsl==5.3.0
-elasticsearch==1.4.0
-Flask-Cors==2.0.1
 google-cloud-pubsub==0.28.4
 google-cloud-storage==1.5.0
-kafka-python==1.3.3
 watchdog==0.8.3
-git+git://github.com/david4096/data-object-schemas@fixes-cleanup
+git+git://github.com/ga4gh/data-object-schemas@master
